Acceleration is now notified as “floats” and changes in some remote status names.

+++ sensing/src/main/java/com/mytechia/robobo/framework/sensing/accel/IAccelerationListener.java
@@ -24,7 +24,7 @@ package com.mytechia.robobo.framework.sensing.accel;
 public interface IAccelerationListener {
 
     void onAccelerationChange();
-    void onAcceleration(int xaccel, int yaccel, int zaccel);
+    void onAcceleration(double xaccel, double yaccel, double zaccel);
     void onCalibrationAngle(double angle);
 
 }
